refactor: Remove extensions from mavlink-kotlin module
