+++ adapters/adapter-coroutines/src/test/kotlin/com/divpundir/mavlink/adapters/coroutines/CoroutinesMavConnectionTest.kt
@@ -1,12 +1,12 @@
 package com.divpundir.mavlink.adapters.coroutines
 
+import com.divpundir.mavlink.api.wrap
 import com.divpundir.mavlink.connection.tcp.TcpClientMavConnection
 import com.divpundir.mavlink.definitions.ardupilotmega.ArdupilotmegaDialect
 import com.divpundir.mavlink.definitions.common.CommandLong
 import com.divpundir.mavlink.definitions.common.CommonDialect
 import com.divpundir.mavlink.definitions.common.MavCmd
 import com.divpundir.mavlink.definitions.common.ScaledImu2
-import com.divpundir.mavlink.wrap
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.filterIsInstance
 import kotlinx.coroutines.flow.map
