+++ buildSrc/src/main/kotlin/Configuration.kt
@@ -5,8 +5,8 @@ object Config {
     val javaVersion = JavaVersion.VERSION_1_8
 
     object Plugin {
-        const val releaseVersion = "0.20.0"
-        const val developmentVersion = "0.30.0-alpha"
+        const val releaseVersion = "0.30.0-alpha"
+        const val developmentVersion = "0.30.0-alpha.2"
     }
 
     object Lib {
